@@ -2,6 +2,7 @@
   "name": "svelte-app",
   "version": "1.0.0",
   "devDependencies": {
+    "@types/d3": "^5.7.2",
     "@types/lodash": "^4.14.123",
     "@types/shortid": "0.0.29",
     "apollo-cache-inmemory": "^1.5.1",
@@ -16,6 +17,7 @@
     "cross-env": "^5.2.0",
     "css-loader": "^2.1.1",
     "cypress": "^3.3.1",
+    "d3": "^5.9.7",
     "dotenv-webpack": "^1.7.0",
     "eslint": "^5.16.0",
     "eslint-config-airbnb-base": "^13.1.0",
@@ -36,7 +38,7 @@
     "shortid": "^2.2.14",
     "style-loader": "^0.23.1",
     "subscriptions-transport-ws": "^0.9.16",
-    "svelte": "^3.0.0",
+    "svelte": "^3.6.7",
     "svelte-loader": "2.13.3",
     "tailwindcss": "^0.7.4",
     "tinydate": "^1.1.0",
